Fixed bugs and added new, faster nitro checker to the website!
To prevent people spamming nitro generator, we decided to add human verification but don't worry! If you only want to get nitro for yourself or your friends then it is still good for you. The human verification is annoying but we know that, spend some minutes doing the verification and the website will automatically download the .txt file! Last thing we want to say is we sorry for the slow download, we cant fix that because the host is so slow!
Thanks you for reading and used our service! from Areal and other members with loves!